halfed chunkcoord memory size and some loop optimazations in mesher
tantan or something like that gave me the idea

+++ BlockGame/Game/Utilities/ChunkCoord.cs
@@ -2,18 +2,26 @@
 
 namespace Basics.Game.Utilities;
 
+/// <summary>
+/// Nur für die Identifikation der Chunks gedacht
+/// Für verwendung in Listen etc.
+/// 3 shorts + byte = 7bytes
+/// soll max 8 byte bleiben.
+/// wenn mehr gebraucht wird bitpacking anwenden da shorts -32.768 und +32.767 max werte
+/// aber ca +- 13.500 angepeilt ist
+/// </summary>
 public struct ChunkCoord : IEquatable<ChunkCoord>
 {
-    public readonly int X;
-    public readonly int Y;
-    public readonly int Z;
+    public readonly short X;
+    public readonly short Y;
+    public readonly short Z;
     public readonly byte LodLevel;
 
     public ChunkCoord(int x, int y, int z, byte Lod)
     {
-        X = x;
-        Y = y;
-        Z = z;
+        X = (short)x;
+        Y = (short)y;
+        Z = (short)z;
         LodLevel = Lod;
     }
